Added damage to projectiles, with basic sounds and ships destruction

+++ src/comp_collision.cpp
@@ -1,6 +1,8 @@
 #include <iostream>
 #include <assert.h>
 
+#include "../include/audiobuffer.h"
+#include "../include/audiosource.h"
 #include "../include/collision.h"
 #include "../include/comp_collision.h"
 #include "../include/comp_render.h"
@@ -20,16 +22,53 @@ void CCompCollision::ReceiveMessage(SMessage & msg) {
 		Sprite * otherSprt = isColMsg.GetOther()->GetRenderComp()->GetSprite();
 		Sprite * sprt = m_owner->GetRenderComp()->GetSprite();
 		if (sprt->CheckCollision(otherSprt)) {
-			if (isColMsg.GetOther()->GetType() == EET_SHIP) {
-				std::cout << "SHIP! BOOOOOOOM!!!" << std::endl;
+			if (m_owner->GetType() == EET_SHIP && isColMsg.GetOther()->GetType() == EET_SHIP) {
+				std::cout << "SHIP - SHIP" << std::endl;
+				AudioBuffer * buffer = new AudioBuffer("data/sounds/explosion3.wav");
+				AudioSource * source = new AudioSource(buffer);
+
 				SGetWorldMsg getWorldMsg;
 				m_owner->ReceiveMessage(getWorldMsg);
 				assert(getWorldMsg.Modified());
 				getWorldMsg.GetWorld()->DeleteEntity(isColMsg.GetOther());
 				getWorldMsg.GetWorld()->DeleteEntity(m_owner);
-			} else if (isColMsg.GetOther()->GetType() == EET_PROJECTILE) {
-				std::cout << "PROJECTILE COLLISION" << std::endl;
-			}
+				source->Play();
+			} else if (m_owner->GetType() == EET_SHIP
+			&& isColMsg.GetOther()->GetType() == EET_PROJECTILE) {
+				std::cout << "SHIP - PROJECTILE" << std::endl;
+				SGetDamageMsg getDmgMsg;
+				isColMsg.GetOther()->ReceiveMessage(getDmgMsg);
+				if (getDmgMsg.Modified()) {
+					SUpdateHitPointsMsg updateHPMsg(-getDmgMsg.GetDamage());
+					m_owner->ReceiveMessage(updateHPMsg);
+					
+					SGetWorldMsg getWorldMsg;
+					m_owner->ReceiveMessage(getWorldMsg);
+					getWorldMsg.GetWorld()->DeleteEntity(isColMsg.GetOther());
+					AudioBuffer * buffer = new AudioBuffer("data/sounds/explosion1.wav");
+					AudioSource * source = new AudioSource(buffer);
+					source->Play();
+				}
+			} else if (m_owner->GetType() == EET_PROJECTILE
+			&& isColMsg.GetOther()->GetType() == EET_SHIP) {
+				std::cout << "PROJECTILE - SHIP" << std::endl;
+				SGetDamageMsg getDmgMsg;
+				m_owner->ReceiveMessage(getDmgMsg);
+				if (getDmgMsg.Modified()) {
+					SUpdateHitPointsMsg updateHPMsg(-getDmgMsg.GetDamage());
+					isColMsg.GetOther()->ReceiveMessage(updateHPMsg);
+
+					SGetWorldMsg getWorldMsg;
+					isColMsg.GetOther()->ReceiveMessage(getWorldMsg);
+					getWorldMsg.GetWorld()->DeleteEntity(m_owner);
+					AudioBuffer * buffer = new AudioBuffer("data/sounds/explosion1.wav");
+					AudioSource * source = new AudioSource(buffer);
+					source->Play();
+				}
+			}/* else if (m_owner->GetType() == EET_PROJECTILE
+			&& isColMsg.GetOther()->GetType() == EET_PROJECTILE) {
+				std::cout << "PROJECTILE - PROJECTILE" << std::endl;
+			}*/
 		}
 	}
 }
